Fixed Cpu_Timer_Mod and renamed it to Profiler_Mod
The reason for these changes was in the fact that "cpu_time",
function which was used in Cpu_Timer_Mod, wasn't giving correct
values when programs (Convert, at this stage) were compiled with
OPENMP=yes.

The version currently submitted, relies on the "system_clock"
instead.  This issue is known and was described here:

https://stackoverflow.com/questions/25465101/fortran-openmp-program-shows-no-speedup-of-cpu-time

This check-in changed a lot of files because all the instances
of Cpu_Timer have been changed to Profiler.

An additional change was done in Convert/Load_Gmsh.  It has
been optimized for reading and is faster than before.

On branch boil_bojan

modified:   Convert/Calculate_Geometry.f90
modified:   Convert/Find_Faces.f90
modified:   Convert/Find_Parents.f90
modified:   Convert/Grid_Topology.f90
modified:   Convert/Load_Fluent.f90
modified:   Convert/Load_Gambit.f90
modified:   Convert/Load_Gmsh.f90
modified:   Convert/Main_Con.f90
modified:   Convert/makefile
modified:   Convert/makefile_explicit_dependencies
modified:   Divide/makefile
modified:   Divide/makefile_explicit_dependencies
modified:   Generate/makefile
modified:   Generate/makefile_explicit_dependencies
modified:   Process/Backup_Mod/Load.f90
modified:   Process/Backup_Mod/Save.f90
modified:   Process/Compute_Energy.f90
modified:   Process/Compute_Momentum.f90
modified:   Process/Compute_Pressure.f90
modified:   Process/Compute_Scalar.f90
modified:   Process/Correct_Velocity.f90
modified:   Process/Field_Mod.f90
modified:   Process/Field_Mod/Gradients/Grad_Gauss_Variable.f90
modified:   Process/Field_Mod/Gradients/Grad_Least_Variable.f90
modified:   Process/Front_Mod.f90
modified:   Process/Front_Mod/Place_Front_At_Value.f90
modified:   Process/Info_Mod/Start.f90
modified:   Process/Info_Mod/Time_Fill.f90
modified:   Process/Interface_Mod.f90
modified:   Process/Main_Pro.f90
modified:   Process/Results_Mod/Save_Results.f90
modified:   Process/Rhie_And_Chow.f90
modified:   Process/Surf_Mod/Place_Surf_At_Value.f90
modified:   Process/Turb_Mod.f90
modified:   Process/Turb_Mod/Compute_F22.f90
modified:   Process/Turb_Mod/Compute_Stress.f90
modified:   Process/Turb_Mod/Compute_Variable.f90
modified:   Process/Vof_Mod/Core/Compute_Vof.f90
modified:   Process/Vof_Mod/Core/Solve_System.f90
modified:   Process/makefile
modified:   Process/makefile_explicit_dependencies
modified:   Shared/Grid_Mod.f90
modified:   Shared/Grid_Mod/Calculate_Wall_Distance.f90
modified:   Shared/Grid_Mod/Save_Cfn.f90
modified:   Shared/Grid_Mod/Save_Dim.f90
modified:   Shared/Grid_Mod/Save_Vtu_Cells.f90
modified:   Shared/Grid_Mod/Save_Vtu_Edges.f90
modified:   Shared/Grid_Mod/Save_Vtu_Faces.f90
renamed:    Shared/Cpu_Timer_Mod.f90 -> Shared/Profiler_Mod.f90
renamed:    Shared/Cpu_Timer_Mod/Start.f90 -> Shared/Profiler_Mod/Start.f90
renamed:    Shared/Cpu_Timer_Mod/Statistics.f90 -> Shared/Profiler_Mod/Statistics.f90
renamed:    Shared/Cpu_Timer_Mod/Stop.f90 -> Shared/Profiler_Mod/Stop.f90
renamed:    Shared/Cpu_Timer_Mod/Update_By_Rank.f90 -> Shared/Profiler_Mod/Update_By_Rank.f90

+++ Sources/Process/Info_Mod/Time_Fill.f90
@@ -8,8 +8,8 @@
   integer :: n         ! time step
   real    :: sim_time  ! simulation time
 !-----------------------------------[Locals]-----------------------------------!
-  integer :: hours, minutes, seconds
-  real    :: wall_time                ! number of seconds of wall-clock time
+  integer  :: hours, minutes, seconds
+  real(DP) :: wall_time                ! number of seconds of wall-clock time
 !==============================================================================!
 
   ! Update current system clock and wall time
